[stabilizer] Compute ZMP in free functions
