+++ src/mc_rbdyn/Robot.cpp
@@ -7,6 +7,7 @@
 #include <mc_rbdyn/Robots.h>
 #include <mc_rbdyn/SCHAddon.h>
 #include <mc_rbdyn/Surface.h>
+#include <mc_rbdyn/ZMP.h>
 #include <mc_rbdyn/constants.h>
 #include <mc_rbdyn/surface_utils.h>
 #include <mc_rtc/logging.h>
@@ -541,32 +542,14 @@ Eigen::Vector3d Robot::zmp(const sva::ForceVecd & netTotalWrench,
                            const Eigen::Vector3d & plane_n,
                            double minimalNetNormalForce) const
 {
-  if(minimalNetNormalForce <= 0)
-  {
-    LOG_ERROR_AND_THROW(std::runtime_error,
-                        "ZMP cannot be computed: the minimalNetNormalForce must be >0 (divide by zero)");
-  }
-
-  const Eigen::Vector3d & force = netTotalWrench.force();
-  const Eigen::Vector3d & moment_0 = netTotalWrench.couple();
-  Eigen::Vector3d moment_p = moment_0 - plane_p.cross(force);
-  double floorn_dot_force = plane_n.dot(force);
-  // Prevent potential division by zero
-  if(floorn_dot_force < minimalNetNormalForce)
-  {
-    LOG_ERROR_AND_THROW(std::runtime_error, "ZMP cannot be computed, projected force too small " << floorn_dot_force);
-  }
-  Eigen::Vector3d zmp = plane_p + plane_n.cross(moment_p) / floorn_dot_force;
-  return zmp;
+  return mc_rbdyn::zmp(netTotalWrench, plane_p, plane_n, minimalNetNormalForce);
 }
 
 Eigen::Vector3d Robot::zmp(const sva::ForceVecd & netWrench,
                            const sva::PTransformd & zmpFrame,
                            double minimalNetNormalForce) const
 {
-  Eigen::Vector3d n = zmpFrame.rotation().row(2);
-  Eigen::Vector3d p = zmpFrame.translation();
-  return zmp(netWrench, p, n, minimalNetNormalForce);
+  return mc_rbdyn::zmp(netWrench, zmpFrame, minimalNetNormalForce);
 }
 
 Eigen::Vector3d Robot::zmp(const std::vector<std::string> & sensorNames,
